+++ backend/src/SSACFG.c
@@ -5,14 +5,14 @@
 void SSACFG_free(SSACFG* cfg) {
     StrBuf_free(&cfg->name);
 
-    for (size_t i = 0; i < cfg->len; ++i) {
+    for (uint32_t i = 0; i < cfg->len; ++i) {
         SSABasicBlock_free(&cfg->blocks[i]);
     }
     mycc_free(cfg->blocks);
 
     mycc_free(cfg->regs);
 
-    for (size_t i = 0; i < cfg->num_val_names; ++i) {
+    for (uint32_t i = 0; i < cfg->num_val_names; ++i) {
         StrBuf_free(&cfg->val_names[i]);
     }
     mycc_free(cfg->val_names);
